% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5.R
\name{KoladaAPI}
\alias{KoladaAPI}
\title{KoladaAPI Main function}
\usage{
KoladaAPI()
}
\description{
KoladaAPI Main function
}
\section{Fields}{

\describe{
\item{\code{noinput}}{}
}}

\author{
Hirbod & Muhaiminul
}
