% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_plots.R
\name{plotfunction_ind}
\alias{plotfunction_ind}
\title{Shinyapp based on KoladaAPI}
\usage{
plotfunction_ind()
}
\description{
Shinyapp based on KoladaAPI
}
\section{Fields}{

\describe{
\item{\code{noinput}}{}
}}

\author{
Hirbod & Muhaiminul
}
